Optimized Stock Listing Added With Pagination Method
Sorry for delay,
Little health issue affected completion time.

Optimized stock listing added with pagination method and also the Object Pooling method is there but commented.

Now at a time only 5 stocks will be shown and on reaching end to the scroll rect it will spawn another 5.

Added Testing.cs that I used to test the optimization logic if its working fine or not and then integrated to the main project.